requirements.txt
altgraph==0.17.2
black==22.8.0
click==8.1.3
dnspython==2.2.1
macholib==1.16
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
